chore: cleaned all duplicate code
