+++ src/hooks/context/index.ts
@@ -1,3 +1,2 @@
 export * from './hooks';
 export * from './models';
-export * from './errors';
